+++ packages/reprint-importer/src/import.php
@@ -38,6 +38,12 @@ require_once __DIR__ . '/lib/mysql-query-stream/load.php';
 // Load WordPress function stubs (needed by wp-php-toolkit outside WordPress)
 require_once __DIR__ . '/lib/wp-stubs.php';
 
+// Streaming protocol parsers.
+require_once __DIR__ . '/lib/protocol/class-multipart-stream-parser.php';
+if (!class_exists('MultipartStreamParser', false)) {
+    class_alias(\Reprint\Importer\Protocol\MultipartStreamParser::class, 'MultipartStreamParser');
+}
+
 // Load URL rewriting components
 require_once __DIR__ . '/lib/url-rewrite/load.php';
 
@@ -230,319 +236,6 @@ function register_sqlite_function(PDO $sqlite_pdo, string $name, callable $fn, i
     }
 }
 
-
-/**
- * Streaming multipart parser.
- * Parses multipart/mixed responses incrementally without buffering entire response.
- */
-class MultipartStreamParser
-{
-    private const MAX_BUFFER_SIZE = 64 * 1024 * 1024; // 64MB
-    private const STATE_BOUNDARY = 0;
-    private const STATE_HEADERS = 1;
-    private const STATE_BODY = 2;
-
-    private $boundary;
-    private $boundary_length;
-    private $buffer = "";
-    private $state = self::STATE_BOUNDARY;
-    private $current_headers = [];
-    private $body_length = 0;
-    private $body_target = null;
-    private $chunk_handler;
-
-    public function __construct(string $boundary, callable $chunk_handler)
-    {
-        $this->boundary = "--" . $boundary;
-        $this->boundary_length = strlen($this->boundary);
-        $this->chunk_handler = $chunk_handler;
-    }
-
-    /**
-     * Feed data to parser. Called by curl write callback.
-     */
-    public function feed(string $data): void
-    {
-        $this->buffer .= $data;
-        if (strlen($this->buffer) > self::MAX_BUFFER_SIZE) {
-            throw new RuntimeException(
-                "Multipart parser buffer exceeded 64MB — response may be malformed (missing boundary delimiter)."
-            );
-        }
-        $this->parse();
-    }
-
-    /**
-     * Parse buffered data.
-     */
-    private function parse(): void
-    {
-        while (true) {
-            if ($this->state === self::STATE_BOUNDARY) {
-                if (!$this->parse_boundary()) {
-                    break;
-                }
-            } elseif ($this->state === self::STATE_HEADERS) {
-                if (!$this->parse_headers()) {
-                    break;
-                }
-            } elseif ($this->state === self::STATE_BODY) {
-                if (!$this->parse_body()) {
-                    break;
-                }
-            }
-        }
-    }
-
-    /**
-     * Parse boundary. Returns true if boundary found and consumed.
-     */
-    private function parse_boundary(): bool
-    {
-        // Look for boundary
-        $pos = strpos($this->buffer, $this->boundary);
-        if ($pos === false) {
-            // Keep only last boundary_length bytes in case boundary is split
-            if (strlen($this->buffer) > $this->boundary_length) {
-                $this->buffer = substr($this->buffer, -$this->boundary_length);
-            }
-            return false;
-        }
-
-        // Check if this is the closing boundary (--boundary--)
-        $after_boundary = $pos + $this->boundary_length;
-        if ($after_boundary + 2 <= strlen($this->buffer)) {
-            $next_chars = substr($this->buffer, $after_boundary, 2);
-            if ($next_chars === "--") {
-                // Closing boundary - done
-                $this->buffer = "";
-                return false;
-            }
-        }
-
-        // Find end of line after boundary (\r\n or \n)
-        $line_end = $this->find_line_end($after_boundary);
-        if ($line_end === false) {
-            return false; // Need more data
-        }
-
-        // Consume boundary line
-        $this->buffer = substr($this->buffer, $line_end);
-        $this->state = self::STATE_HEADERS;
-        $this->current_headers = [];
-        return true;
-    }
-
-    /**
-     * Parse headers. Returns true if all headers parsed.
-     */
-    private function parse_headers(): bool
-    {
-        while (true) {
-            // Check for blank line (end of headers)
-            if (strlen($this->buffer) >= 2) {
-                if ($this->buffer[0] === "\r" && $this->buffer[1] === "\n") {
-                    // \r\n - blank line
-                    $this->buffer = substr($this->buffer, 2);
-                    $this->prepare_body();
-                    return true;
-                } elseif ($this->buffer[0] === "\n") {
-                    // \n - blank line
-                    $this->buffer = substr($this->buffer, 1);
-                    $this->prepare_body();
-                    return true;
-                }
-            }
-
-            // Find end of line
-            $line_end = $this->find_line_end(0);
-            if ($line_end === false) {
-                return false; // Need more data
-            }
-
-            // Extract header line
-            $line = substr($this->buffer, 0, $line_end);
-            $this->buffer = substr($this->buffer, $line_end);
-
-            // Trim line endings
-            $line = rtrim($line, "\r\n");
-
-            if ($line === "") {
-                // Blank line - end of headers
-                $this->prepare_body();
-                return true;
-            }
-
-            // Parse header (find first colon)
-            $colon_pos = strpos($line, ":");
-            if ($colon_pos !== false) {
-                $name = substr($line, 0, $colon_pos);
-                $value = substr($line, $colon_pos + 1);
-
-                // Trim spaces
-                $name = trim($name);
-                $value = ltrim($value); // Only left trim value
-
-                // Store header (lowercase key)
-                $key = strtolower($name);
-                $this->current_headers[$key] = $value;
-            }
-        }
-    }
-
-    /**
-     * Prepare for body parsing.
-     */
-    private function prepare_body(): void
-    {
-        $this->state = self::STATE_BODY;
-        $this->body_length = 0;
-
-        // Determine target length if Content-Length is specified
-        $this->body_target = isset($this->current_headers["content-length"])
-            ? (int) $this->current_headers["content-length"]
-            : null;
-    }
-
-    /**
-     * Parse body. Returns true if body complete.
-     */
-    private function parse_body(): bool
-    {
-        // If we know the content length, read exactly that many bytes
-        if ($this->body_target !== null) {
-            $remaining = $this->body_target - $this->body_length;
-
-            if (strlen($this->buffer) < $remaining) {
-                // Need more data
-                if (strlen($this->buffer) > 0) {
-                    // Process what we have
-                    $this->emit_body_chunk(substr($this->buffer, 0));
-                    $this->body_length += strlen($this->buffer);
-                    $this->buffer = "";
-                }
-                return false;
-            }
-
-            // We have enough data
-            $body_data = substr($this->buffer, 0, $remaining);
-            $this->buffer = substr($this->buffer, $remaining);
-
-            $this->emit_body_chunk($body_data);
-            $this->body_length += strlen($body_data);
-
-            // Skip trailing \r\n after body
-            $this->skip_crlf();
-
-            // Complete - move to next boundary
-            $this->state = self::STATE_BOUNDARY;
-            $this->emit_chunk_complete();
-            return true;
-        }
-
-        // No content-length - read until next boundary
-        // Look for boundary in buffer
-        $boundary_pos = strpos($this->buffer, "\r\n" . $this->boundary);
-        if ($boundary_pos === false) {
-            $boundary_pos = strpos($this->buffer, "\n" . $this->boundary);
-        }
-
-        if ($boundary_pos === false) {
-            // No boundary yet - process all but last boundary_length+2 bytes
-            $safe_length = strlen($this->buffer) - $this->boundary_length - 2;
-            if ($safe_length > 0) {
-                $body_data = substr($this->buffer, 0, $safe_length);
-                $this->buffer = substr($this->buffer, $safe_length);
-                $this->emit_body_chunk($body_data);
-                $this->body_length += strlen($body_data);
-            }
-            return false;
-        }
-
-        // Found boundary - emit remaining body
-        $body_data = substr($this->buffer, 0, $boundary_pos);
-        $this->buffer = substr($this->buffer, $boundary_pos);
-
-        $this->emit_body_chunk($body_data);
-        $this->body_length += strlen($body_data);
-
-        // Skip \r\n before boundary
-        $this->skip_crlf();
-
-        // Complete - move to next boundary
-        $this->state = self::STATE_BOUNDARY;
-        $this->emit_chunk_complete();
-        return true;
-    }
-
-    /**
-     * Skip \r\n or \n at start of buffer.
-     */
-    private function skip_crlf(): void
-    {
-        if (
-            strlen($this->buffer) >= 2 &&
-            $this->buffer[0] === "\r" &&
-            $this->buffer[1] === "\n"
-        ) {
-            $this->buffer = substr($this->buffer, 2);
-        } elseif (strlen($this->buffer) >= 1 && $this->buffer[0] === "\n") {
-            $this->buffer = substr($this->buffer, 1);
-        }
-    }
-
-    /**
-     * Find line end position (\r\n or \n) starting from offset.
-     * Returns position after line ending, or false if not found.
-     */
-    /** @return int|false */
-    private function find_line_end(int $offset)
-    {
-        $len = strlen($this->buffer);
-
-        for ($i = $offset; $i < $len; $i++) {
-            if ($this->buffer[$i] === "\n") {
-                return $i + 1;
-            }
-            if (
-                $this->buffer[$i] === "\r" &&
-                $i + 1 < $len &&
-                $this->buffer[$i + 1] === "\n"
-            ) {
-                return $i + 2;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Emit body chunk to handler.
-     */
-    private function emit_body_chunk(string $data): void
-    {
-        if ($data === "") {
-            return;
-        }
-
-        ($this->chunk_handler)([
-            "type" => "body",
-            "headers" => $this->current_headers,
-            "data" => $data,
-        ]);
-    }
-
-    /**
-     * Emit chunk complete to handler.
-     */
-    private function emit_chunk_complete(): void
-    {
-        ($this->chunk_handler)([
-            "type" => "complete",
-            "headers" => $this->current_headers,
-        ]);
-    }
-}
 
 /**
  * AdaptiveTuner
@@ -11217,7 +10910,7 @@ class ImportClient
                                 "Creating multipart parser with boundary: $boundary_value",
                                 false,
                             );
-                            $parser = new MultipartStreamParser(
+                            $parser = new \Reprint\Importer\Protocol\MultipartStreamParser(
                                 $boundary_value,
                                 $this->make_chunk_handler($context, $current_chunk),
                             );
@@ -11256,7 +10949,7 @@ class ImportClient
                                         "Detected boundary in body (no Content-Type): {$boundary}",
                                         false,
                                     );
-                                    $parser = new MultipartStreamParser(
+                                    $parser = new \Reprint\Importer\Protocol\MultipartStreamParser(
                                         $boundary,
                                         $this->make_chunk_handler($context, $current_chunk),
                                     );
